Refactor TUI model to export struct and methods; reorganize CLI and introduce `usb-tree` entry point
